+++ config.py
@@ -1,0 +1,8 @@
+IP_CAMERA_URL = "http://192.168.1.66:8080/video"  # URL of your IP camera
+KNOWN_FACES_DATA = [
+    ("Pessoa 1", "Morty.png"),
+    # ("Person Name 2", "path/to/image2.jpg"),
+]
+RESIZE_FACTOR = 0.5
+TOLERANCE = 0.6
+OUTPUT_DIR = "detected_faces_output"
